+++ internal/service/encryptionatrest/model_test.go
@@ -3,7 +3,7 @@ package encryptionatrest_test
 import (
 	"testing"
 
-	"go.mongodb.org/atlas-sdk/v20250219001/admin"
+	"go.mongodb.org/atlas-sdk/v20250312002/admin"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/stretchr/testify/assert"
@@ -83,9 +83,10 @@ var (
 		ServiceAccountKey:    types.StringValue(serviceAccountKey),
 	}
 	EncryptionAtRest = &admin.EncryptionAtRest{
-		AwsKms:         AWSKMSConfiguration,
-		AzureKeyVault:  AzureKeyVault,
-		GoogleCloudKms: GoogleCloudKMS,
+		AwsKms:                AWSKMSConfiguration,
+		AzureKeyVault:         AzureKeyVault,
+		GoogleCloudKms:        GoogleCloudKMS,
+		EnabledForSearchNodes: &enabled,
 	}
 )
 
@@ -99,11 +100,12 @@ func TestNewTfEncryptionAtRestRSModel(t *testing.T) {
 			name:     "Success NewTFAwsKmsConfig",
 			sdkModel: EncryptionAtRest,
 			expectedResult: &encryptionatrest.TfEncryptionAtRestRSModel{
-				ID:                   types.StringValue(projectID),
-				ProjectID:            types.StringValue(projectID),
-				AwsKmsConfig:         []encryptionatrest.TFAwsKmsConfigModel{TfAwsKmsConfigModel},
-				AzureKeyVaultConfig:  []encryptionatrest.TFAzureKeyVaultConfigModel{TfAzureKeyVaultConfigModel},
-				GoogleCloudKmsConfig: []encryptionatrest.TFGcpKmsConfigModel{TfGcpKmsConfigModel},
+				ID:                    types.StringValue(projectID),
+				ProjectID:             types.StringValue(projectID),
+				AwsKmsConfig:          []encryptionatrest.TFAwsKmsConfigModel{TfAwsKmsConfigModel},
+				AzureKeyVaultConfig:   []encryptionatrest.TFAzureKeyVaultConfigModel{TfAzureKeyVaultConfigModel},
+				GoogleCloudKmsConfig:  []encryptionatrest.TFGcpKmsConfigModel{TfGcpKmsConfigModel},
+				EnabledForSearchNodes: types.BoolValue(enabled),
 			},
 		},
 	}